time: Portability modifications
This commit addresses the following portability issues:

- native integer (pointer) types being cast directly to a 64 bit
  Duration type.

- constant integer division (1/60e90) does not result in a double where
  it was expected to be.

- fn_compare was casting bool results directly to int which broke on
  MacOS-Arm64.

- Use HAVE_TIMESPEC_GET feature detection to leave out unsupported code
  and tests. This enables most of the time extension to run on 32bit
  Windows.

+++ test/time/duration.test.c
@@ -116,7 +116,7 @@ typedef struct {
 } ToMinuteTest;
 
 static ToMinuteTest to_minute_tests[] = {
-    {-60000000000, -1}, {-1, -1 / 60e9}, {1, 1 / 60e9}, {60000000000, 1}, {3000, 5e-8},
+    {-60000000000, -1}, {-1, (double)-1 / 60e9}, {1, (double)1 / 60e9}, {60000000000, 1}, {3000, (double)5e-8},
 };
 
 static void test_to_minutes(void) {
@@ -124,7 +124,7 @@ static void test_to_minutes(void) {
     for (size_t i = 0; i < sizeof(to_minute_tests) / sizeof(to_minute_tests[0]); i++) {
         ToMinuteTest test = to_minute_tests[i];
         double got = dur_to_minutes(test.d);
-        // printf("want %f, got %f\n", test.want, got);
+        // printf("want %lf, got %lf\n", test.want, got);
         assert(got == test.want);
     }
     printf("OK\n");
@@ -136,7 +136,7 @@ typedef struct {
 } ToHourTest;
 
 static ToHourTest to_hour_tests[] = {
-    {-3600000000000, -1}, {-1, -1 / 3600e9}, {1, 1 / 3600e9}, {3600000000000, 1}, {36, 1e-11},
+    {-3600000000000, -1}, {-1, (double)-1 / 3600e9}, {1, (double)1 / 3600e9}, {3600000000000, 1}, {36, (double)1e-11},
 };
 
 static void test_to_hours(void) {
@@ -144,7 +144,7 @@ static void test_to_hours(void) {
     for (size_t i = 0; i < sizeof(to_hour_tests) / sizeof(to_hour_tests[0]); i++) {
         ToHourTest test = to_hour_tests[i];
         double got = dur_to_hours(test.d);
-        // printf("want %f, got %f\n", test.want, got);
+        // printf("want %lf, got %lf\n", test.want, got);
         assert(got == test.want);
     }
     printf("OK\n");
